+++ src/main/java/delayCalculator/DelayCalculator.java
@@ -3,6 +3,8 @@ package delayCalculator;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import delayCalculator.delayOptions.DelayOptions;
+
 /**
  * TODO: finish javadocs
  */
