+++ src/main/java/delayCalculator/delayOptions/DelayOptions.java
@@ -1,4 +1,4 @@
-package delayCalculator;
+package delayCalculator.delayOptions;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
